edm_web/requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
chardet==2.3.0
Django==1.10.1
django-redis==4.6.0
django-user-agents==0.3.0
django-simple-captcha==0.5.6
dnspython==1.15.0
gevent==1.1.2
greenlet==0.4.10
image==1.5.5
#mod-wsgi==4.5.3
MySQL-python==1.2.5
passlib==1.7.0
phpserialize==1.3
Pillow==3.4.2
pycrypto==2.6.1
pymongo==3.4.0
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
qrcode==5.3
redis==2.10.5
requests==2.6.0
six==1.10.0
ua-parser==0.7.2
urllib3==1.19.1
user-agents==1.0.1
wechat-sdk==0.6.4
xlwt==1.1.2
xmltodict==0.9.2
django-jsonify==0.3.0
django-jsonfield==1.0.1
sorl-thumbnail==12.3
premailer==3.1.1
django-debug-toolbar==1.8
django-wkhtmltopdf==3.1.0
xhtml2pdf==0.0.6